+++ src/MoBi.Presentation/Presenter/ConfigureSimulationPresenter.cs
@@ -1,20 +1,20 @@
 using MoBi.Assets;
-using OSPSuite.Core.Services;
-using OSPSuite.Utility.Extensions;
 using MoBi.Core.Commands;
 using MoBi.Core.Domain.Model;
 using MoBi.Core.Domain.Model.Diagram;
 using MoBi.Core.Services;
 using MoBi.Presentation.Presenter.Simulation;
 using OSPSuite.Core.Domain.Builder;
+using OSPSuite.Core.Services;
 using OSPSuite.Presentation.Core;
 using OSPSuite.Presentation.Presenters;
+using OSPSuite.Utility.Extensions;
 
 namespace MoBi.Presentation.Presenter
 {
    public interface IConfigureSimulationPresenter : IWizardPresenter
    {
-      IMoBiCommand CreateBuildConfigurationBaseOn(IMoBiSimulation simulation, IBuildingBlock templateBuildingBlock);
+      IMoBiCommand CreateBuildConfigurationBasedOn(IMoBiSimulation simulation, IBuildingBlock templateBuildingBlock);
       IMoBiCommand CreateBuildConfiguration(IMoBiSimulation simulation);
       SimulationConfiguration SimulationConfiguration { get; }
    }
@@ -24,24 +24,27 @@ namespace MoBi.Presentation.Presenter
       private readonly IDiagramManagerFactory _diagramManagerFactory;
       public SimulationConfiguration SimulationConfiguration { get; private set; }
 
-
       public ConfigureSimulationPresenter(IConfigureSimulationView view, ISubPresenterItemManager<ISimulationConfigurationItemPresenter> subPresenterSubjectManager, IDialogCreator dialogCreator, IBuildConfigurationFactory buildConfigurationFactory, IMoBiContext context, IDiagramManagerFactory diagramManagerFactory)
          : base(view, subPresenterSubjectManager, dialogCreator, buildConfigurationFactory, context, SimulationItems.AllConfigure)
       {
          _diagramManagerFactory = diagramManagerFactory;
       }
 
-      public IMoBiCommand CreateBuildConfiguration(IMoBiSimulation simulation) => CreateBuildConfigurationBaseOn(simulation, null);
+      public IMoBiCommand CreateBuildConfiguration(IMoBiSimulation simulation) => CreateBuildConfigurationBasedOn(simulation, null);
 
-
-      public IMoBiCommand CreateBuildConfigurationBaseOn(IMoBiSimulation simulation, IBuildingBlock templateBuildingBlock)
+      public IMoBiCommand CreateBuildConfigurationBasedOn(IMoBiSimulation simulation, IBuildingBlock templateBuildingBlock)
       {
          //we create a build configuration where all current building blocks are referencing template building blocks
+         // SimulationConfiguration = _buildConfigurationFactory.CreateFromReferencesUsedIn(simulation.Configuration, templateBuildingBlock);
+
+         // TODO should this be a clone?
+         SimulationConfiguration = simulation.Configuration;
+
          var tmpSimulation = new MoBiSimulation()
          {
             DiagramManager = _diagramManagerFactory.Create<ISimulationDiagramManager>(),
-            // TODO should this be a clone?
-            Configuration = simulation.Configuration,
+
+            Configuration = SimulationConfiguration,
             Creation = simulation.Creation,
             Name = simulation.Name,
          };
