+++ src/MoBi.Presentation/Views/IBuildingBlockSelectionView.cs
@@ -4,10 +4,6 @@ using OSPSuite.Presentation.Views;
 
 namespace MoBi.Presentation.Views
 {
-   public interface IModuleSelectionView : IView<IModuleSelectionPresenter>, IResizableView
-   {
-      void BindTo(ModuleSelectionDTO moduleSelectionDTO);
-   }
    public interface IBuildingBlockSelectionView : IView<IBuildingBlockSelectionPresenter>, IResizableView
    {
       bool NewVisible { set; }
