+++ src/MoBi.UI/Views/BuildingBlockSelectionView.cs
@@ -10,7 +10,7 @@ namespace MoBi.UI.Views
    {
       private ScreenBinder<BuildingBlockSelectionDTO> _screenBinder;
       private IBuildingBlockSelectionPresenter _presenter;
-      
+
       public override void InitializeBinding()
       {
          _screenBinder = new ScreenBinder<BuildingBlockSelectionDTO>();
@@ -25,12 +25,12 @@ namespace MoBi.UI.Views
       }
 
       public override bool HasError => _screenBinder.HasError;
-      
+
       public void AttachPresenter(IBuildingBlockSelectionPresenter presenter)
       {
          _presenter = presenter;
       }
-      
+
       public void BindTo(BuildingBlockSelectionDTO buildingBlockSelectionDTO)
       {
          _screenBinder.BindToSource(buildingBlockSelectionDTO);
@@ -45,40 +45,6 @@ namespace MoBi.UI.Views
       protected override void DisposeBinders()
       {
          _screenBinder.Dispose();
-      }
-   }
-
-   public class ModuleSelectionView : ObjectBaseSelectionView, IModuleSelectionView
-   {
-      private IModuleSelectionPresenter _presenter;
-      private ScreenBinder<ModuleSelectionDTO> _screenBinder;
-      public override void InitializeBinding()
-      {
-         _screenBinder = new ScreenBinder<ModuleSelectionDTO>();
-         _screenBinder.Bind(x => x.SelectedObject)
-            .To(cbBuildingBlocks)
-            .WithValues(x => _presenter.AllAvailableModules)
-            .AndDisplays(x => _presenter.DisplayNameFor(x))
-            .Changed += () => OnEvent(_presenter.SelectedModuleChanged);
-
-         RegisterValidationFor(_screenBinder, NotifyViewChanged);
-         btnNew.Click += (o, e) => OnEvent(() => _presenter.CreateNew());
-      }
-      
-      protected override void DisposeBinders()
-      {
-         _screenBinder.Dispose();
-      }
-
-      public void AttachPresenter(IModuleSelectionPresenter presenter)
-      {
-         _presenter = presenter;
-      }
-
-      public void BindTo(ModuleSelectionDTO moduleSelectionDTO)
-      {
-         _screenBinder.BindToSource(moduleSelectionDTO);
-         AdjustHeight();
       }
    }
 }
